requirements.txt
# Core trading dependencies
ccxt>=3.0.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI and Machine Learning (Phase 3)
tensorflow-cpu>=2.15.0
opencv-python-headless>=4.8.0

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Performance optimization
cython>=0.29.0

# Time and date handling
python-dateutil>=2.8.0

# JSON and data handling
ujson>=5.8.0

# HTTP client for API calls
httpx>=0.24.0

# Websocket support for real-time data
websocket-client>=1.6.0

# Async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Memory optimization
psutil>=5.9.0

# Logging enhancements
colorlog>=6.7.0

# Configuration management
pyyaml>=6.0

# Encryption (for secure API key storage)
cryptography>=41.0.0

